Поправил замечания по 3 спринту (работа с socket - отправка, получение сообщений)
